move event listener to component factory
